feat(term): use uintptr instead of int
Simplify API usage

+++ term/term.go
@@ -6,44 +6,44 @@ type State struct {
 }
 
 // IsTerminal returns whether the given file descriptor is a terminal.
-func IsTerminal(fd int) bool {
+func IsTerminal(fd uintptr) bool {
 	return isTerminal(fd)
 }
 
 // MakeRaw puts the terminal connected to the given file descriptor into raw
 // mode and returns the previous state of the terminal so that it can be
 // restored.
-func MakeRaw(fd int) (*State, error) {
+func MakeRaw(fd uintptr) (*State, error) {
 	return makeRaw(fd)
 }
 
 // GetState returns the current state of a terminal which may be useful to
 // restore the terminal after a signal.
-func GetState(fd int) (*State, error) {
+func GetState(fd uintptr) (*State, error) {
 	return getState(fd)
 }
 
 // SetState sets the given state of the terminal.
-func SetState(fd int, state *State) error {
+func SetState(fd uintptr, state *State) error {
 	return setState(fd, state)
 }
 
 // Restore restores the terminal connected to the given file descriptor to a
 // previous state.
-func Restore(fd int, oldState *State) error {
+func Restore(fd uintptr, oldState *State) error {
 	return restore(fd, oldState)
 }
 
 // GetSize returns the visible dimensions of the given terminal.
 //
 // These dimensions don't include any scrollback buffer height.
-func GetSize(fd int) (width, height int, err error) {
+func GetSize(fd uintptr) (width, height int, err error) {
 	return getSize(fd)
 }
 
 // ReadPassword reads a line of input from a terminal without local echo.  This
 // is commonly used for inputting passwords and other sensitive data. The slice
 // returned does not include the \n.
-func ReadPassword(fd int) ([]byte, error) {
+func ReadPassword(fd uintptr) ([]byte, error) {
 	return readPassword(fd)
 }
