+++ term/term_other.go
@@ -10,30 +10,30 @@ import (
 
 type state struct{}
 
-func isTerminal(fd int) bool {
+func isTerminal(fd uintptr) bool {
 	return false
 }
 
-func makeRaw(fd int) (*State, error) {
+func makeRaw(fd uintptr) (*State, error) {
 	return nil, fmt.Errorf("terminal: MakeRaw not implemented on %s/%s", runtime.GOOS, runtime.GOARCH)
 }
 
-func getState(fd int) (*State, error) {
+func getState(fd uintptr) (*State, error) {
 	return nil, fmt.Errorf("terminal: GetState not implemented on %s/%s", runtime.GOOS, runtime.GOARCH)
 }
 
-func restore(fd int, state *State) error {
+func restore(fd uintptr, state *State) error {
 	return fmt.Errorf("terminal: Restore not implemented on %s/%s", runtime.GOOS, runtime.GOARCH)
 }
 
-func getSize(fd int) (width, height int, err error) {
+func getSize(fd uintptr) (width, height int, err error) {
 	return 0, 0, fmt.Errorf("terminal: GetSize not implemented on %s/%s", runtime.GOOS, runtime.GOARCH)
 }
 
-func setState(fd int, state *State) error {
+func setState(fd uintptr, state *State) error {
 	return fmt.Errorf("terminal: SetState not implemented on %s/%s", runtime.GOOS, runtime.GOARCH)
 }
 
-func readPassword(fd int) ([]byte, error) {
+func readPassword(fd uintptr) ([]byte, error) {
 	return nil, fmt.Errorf("terminal: ReadPassword not implemented on %s/%s", runtime.GOOS, runtime.GOARCH)
 }
